requirements.txt
jupyter 
matplotlib
plotly

scikit-learn
rapidfuzz

nltk
spacy 
tokenizers
textblob

openai
python-dotenv
